books/pyproject.toml
[tool.poetry]
name = "books"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.9"
strawberry-graphql = "^0.73.4"
starlette = "^0.16.0"
uvicorn = "^0.13.4"
taskipy = "^1.8.1"

[tool.taskipy.tasks]
server = "uvicorn app:app --host 0.0.0.0 --port 8000 --reload"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
